Fixed IO restrictions and small refactor
